requirements.txt
Pillow
ansicolors
